feat(model-ad): add marmoset model details e2e tests (MG-1017)
